+++ learning_paths/compat.py
@@ -2,6 +2,8 @@
 Compatibility layer for testing without Open edX.
 """
 
+from datetime import datetime
+
 from django.contrib.auth.models import AbstractBaseUser
 from opaque_keys.edx.keys import CourseKey
 
@@ -41,8 +43,23 @@ def get_course_keys_with_outlines():
     return course_keys_with_outlines()
 
 
+def get_course_due_date(course_key: CourseKey) -> datetime | None:
+    """
+    Retrieve course end date.
+    """
+    # pylint: disable=import-outside-toplevel, import-error
+    from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
+
+    try:
+        overview = CourseOverview.objects.get(id=course_key)
+        return overview.end
+    except CourseOverview.DoesNotExist:
+        return None
+
+
 __all__ = [
     "get_course_keys_with_outlines",
     "get_catalog_api_client",
     "get_user_course_grade",
+    "get_course_due_date",
 ]